# docs/openapi.yaml
openapi: "3.1.0"
info:
  title: Lebellion Management API
  description: |
    Checklist & KPI inspection system for restaurant chains.

    ## Phase 1 — authentication & roles
    - Organization registration by a Founder.
    - Founder / Branch Manager login by email OR phone + password.
    - Employee / Branch Manager join by a one-time, TTL-limited invite code bound to an
      organization and a branch (and a unit for employees).
    - JWT access token (~15 min) + rotating refresh token with theft detection.

    ## Versioning & client gating
    - All endpoints live under `/api/v1` (except `/health`, which is unversioned for infra
      monitoring and is exempt from client-version gating).
    - Every `/api/v1` request MUST send the `X-App-Version` header (semver of the client app).
      Requests from an unsupported/too-old client are rejected with `426 UPGRADE_REQUIRED`
      before any business logic runs. The minimum supported version is a server config value.

    ## Security notes reflected in this contract
    - Passwords are never returned or logged. Minimum length is 10 characters.
    - **Registration gating:** `POST /auth/register` is disabled by default and controlled by
      the `REGISTRATION_ENABLED` env flag; self-serve onboarding is opened deliberately. When
      disabled it returns `403 REGISTRATION_DISABLED`.
    - Access tokens are short-lived JWTs; refresh tokens are opaque, high-entropy strings.
...
      `401 INVALID_TOKEN`.
...

...
  version: "0.4.0"
...
servers:
...
paths:
...
  /api/v1/employees:
    parameters:
      - $ref: "#/components/parameters/XAppVersion"
    get:
      summary: List employees (paginated)
      description: |
        FOUNDER sees all employees in the org; BRANCH_MANAGER sees only employees of their own
        branch. Optional `branchId`/`unitId` filters are further constrained by the caller's scope.
      operationId: listEmployees
      tags:
        - Employees
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Cursor"
        - name: branchId
          in: query
          required: false
          schema:
            type: string
            format: uuid
        - name: unitId
          in: query
          required: false
          schema:
            type: string
            format: uuid
        - name: active
          in: query
          required: false
          schema:
            type: boolean
      responses:
        "200":
          description: A page of employees
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmployeePage"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "426":
          $ref: "#/components/responses/UpgradeRequired"
  /api/v1/employees/{employeeId}/deactivate:
    parameters:
      - $ref: "#/components/parameters/XAppVersion"
      - name: employeeId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    post:
      summary: Deactivate an employee and revoke all their tokens
      description: |
        Marks the user inactive, bumps their `tokenVersion` (so outstanding ACCESS tokens are
        rejected on the very next request, not up to ~15 min later) and revokes ALL of their
        refresh-token families (immediate sign-out everywhere).
        FOUNDER may deactivate any BRANCH_MANAGER/EMPLOYEE in the org;
        BRANCH_MANAGER may deactivate EMPLOYEEs of their OWN branch only. A user cannot
        deactivate themselves, and a FOUNDER cannot be deactivated via this endpoint.
      operationId: deactivateEmployee
      tags:
        - Employees
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Employee deactivated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          description: Employee not found in the caller's scope
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: Target cannot be deactivated (self, or a FOUNDER)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "426":
          $ref: "#/components/responses/UpgradeRequired"
  /api/v1/employees/{employeeId}/recovery-invite:
    parameters:
      - $ref: "#/components/parameters/XAppVersion"
      - name: employeeId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    post:
      summary: Issue a recovery invite (re-bind a lost/replaced device; manager password reset)
      description: |
        Mints a one-time recovery code that re-binds an EXISTING user to a new device at join. Join
        is passwordless in both cases — it revokes the user's old sessions and issues a fresh one.
        Roles: for an EMPLOYEE target the initiator is the FOUNDER or the BRANCH_MANAGER of that
        employee's own branch; for a BRANCH_MANAGER target the initiator is the FOUNDER only.
        A FOUNDER cannot be recovered here (`409`; recovery of the sole founder is an admin-CLI
        procedure — see docs/RUNBOOK.md). An inactive target is `409`.

        Behaviour differs by target role, applied immediately at creation:
        - EMPLOYEE (typically a lost/replaced phone, not theft): only the invite is created; the
          user's existing sessions are revoked later, AT JOIN. No immediate lockout, so an honest
          employee keeps access until they re-bind.
        - BRANCH_MANAGER (typically a password compromise): a temporary password is generated and
          stored with `mustChangePassword=true`, and ALL the manager's sessions are revoked with
          their `tokenVersion` bumped RIGHT AWAY — the old/compromised session dies immediately, not
          at join. The temporary password is returned once (for the Founder to relay) and lets the
          manager log in via /login from another device before changing it.
        Minting a code is rate limited per caller (`429`).
      operationId: issueRecoveryInvite
      tags:
        - Employees
      security:
        - bearerAuth: []
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RecoveryInviteRequest"
      responses:
        "201":
          description: Recovery invite created (code — and, for a manager, a temporary password — returned once)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecoveryInviteResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          description: Target not found in the caller's scope
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: Target cannot be recovered via this endpoint (a FOUNDER, or an inactive user)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "426":
          $ref: "#/components/responses/UpgradeRequired"
        "429":
          $ref: "#/components/responses/RateLimited"
  /api/v1/employees/{employeeId}/permissions:
    parameters:
      - $ref: "#/components/parameters/XAppVersion"
      - name: employeeId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    put:
      summary: Grant/revoke can_score & can_review on an employee (FOUNDER only)
      description: |
        Sets the granular review/score grants (full desired state). FOUNDER-only — a FOUNDER already
        holds both implicitly, and a unit lead can review regardless. The flags are read from the DB on
        each request, so a change takes effect immediately without re-issuing the user's token.
      operationId: setEmployeePermissions
      tags:
        - Employees
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetPermissionsRequest"
      responses:
        "200":
          description: The updated employee profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          description: Employee not found in the caller's scope
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "426":
          $ref: "#/components/responses/UpgradeRequired"
...
  /api/v1/task-instances/{taskInstanceId}/manual-score:
    parameters:
      - $ref: "#/components/parameters/XAppVersion"
      - name: taskInstanceId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    post:
      summary: Score a MANUAL task (FULL/PARTIAL/ZERO)
      description: |
        Requires `can_score` (FOUNDER or an `app_user.can_score` holder); a scorer may not score a task
        they submitted or helped on (`403 CANNOT_SCORE_OWN`). Awards `points × {1, 0.5, 0}`. One score per
        task (re-scoring updates it). Recommendation input only — the system never computes pay.
      operationId: scoreManual
      tags:
        - Review
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ManualScoreRequest"
      responses:
        "200":
          description: The recorded score
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskScore"
        "400":
          description: "`NOT_SCORABLE` — the item is not MANUAL"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          description: "`FORBIDDEN` (no can_score) or `CANNOT_SCORE_OWN`"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Task not found in the caller's scope
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "426":
          $ref: "#/components/responses/UpgradeRequired"
  /api/v1/task-instances/{taskInstanceId}/numeric-score:
    parameters:
      - $ref: "#/components/parameters/XAppVersion"
      - name: taskInstanceId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    post:
      summary: Score a NUMERIC task (value → band → points)
      description: |
        Requires `can_score`; no self-scoring. The value is mapped through the item's bands
        (`[lower, upper)`) to points; a value with no matching band is `400 NOT_SCORABLE`.
      operationId: scoreNumeric
      tags:
        - Review
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NumericScoreRequest"
      responses:
        "200":
          description: The recorded score
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskScore"
        "400":
          description: "`NOT_SCORABLE` — the item is not NUMERIC, or no band matches the value"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          description: Task not found in the caller's scope
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "426":
          $ref: "#/components/responses/UpgradeRequired"
  /api/v1/items/{itemId}/numeric-bands:
    parameters:
      - $ref: "#/components/parameters/XAppVersion"
      - name: itemId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      summary: List a NUMERIC item's scoring bands
      operationId: listNumericBands
      tags:
        - Review
      security:
        - bearerAuth: []
      responses:
        "200":
          description: The item's bands
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/NumericBand"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          description: Item not found in the caller's scope
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "426":
          $ref: "#/components/responses/UpgradeRequired"
    post:
      summary: Add a scoring band to a NUMERIC item (FOUNDER only)
      description: Band is `[lowerBound, upperBound)` — null lower = -inf, null upper = +inf.
      operationId: createNumericBand
      tags:
        - Review
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateNumericBandRequest"
      responses:
        "201":
          description: Band created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NumericBand"
        "400":
          description: "`NOT_SCORABLE` — not a NUMERIC item, or lowerBound >= upperBound"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          description: Item not found in the caller's scope
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "426":
          $ref: "#/components/responses/UpgradeRequired"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Short-lived JWT access token issued by the /auth endpoints.
  parameters:
    XAppVersion:
      name: X-App-Version
      in: header
      required: true
      description: |
        Client app version (semver, e.g. `1.4.0`). Required on all `/api/v1` requests.
        Too-old clients get `426 UPGRADE_REQUIRED`.
      schema:
        type: string
        example: "1.4.0"
...
    Limit:
      name: limit
      in: query
      required: false
      description: Max items per page.
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    Cursor:
      name: cursor
      in: query
      required: false
      description: Opaque cursor from a previous page's `nextCursor`. Omit for the first page.
      schema:
        type: string
  responses:
    BadRequest:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: Missing, expired, forged or otherwise invalid access token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    Forbidden:
      description: |
        Authenticated but not allowed to perform this action. Either the role/branch scope forbids it
        (code `FORBIDDEN`), or the user still has `mustChangePassword` set and must change their password
        before calling any endpoint other than `POST /api/v1/auth/change-password` (code `MUST_CHANGE_PASSWORD`).
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          examples:
            forbidden:
              value:
                code: FORBIDDEN
                message: Access denied
            mustChangePassword:
              value:
                code: MUST_CHANGE_PASSWORD
                message: Password change required
    UpgradeRequired:
      description: Client app version is no longer supported; upgrade required.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            code: UPGRADE_REQUIRED
            message: Please update the app to continue
    RateLimited:
      description: Too many attempts; retry later
      headers:
        Retry-After:
          description: Seconds to wait before retrying
          schema:
            type: integer
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            code: RATE_LIMITED
            message: Too many requests
  schemas:
...
    Role:
      type: string
      enum:
        - FOUNDER
        - BRANCH_MANAGER
        - EMPLOYEE
    InvitableRole:
      type: string
      description: Roles that can be granted through an invite (FOUNDER is never invitable).
      enum:
        - BRANCH_MANAGER
        - EMPLOYEE
...
    UserProfile:
      type: object
      required:
        - id
        - organizationId
        - fullName
        - role
        - active
        - mustChangePassword
        - canScore
        - canReview
      properties:
        id:
          type: string
          format: uuid
        organizationId:
          type: string
          format: uuid
        fullName:
          type: string
        role:
          $ref: "#/components/schemas/Role"
        active:
          type: boolean
        email:
          type: [string, "null"]
          format: email
        phone:
          type: [string, "null"]
        branchId:
          type: [string, "null"]
          format: uuid
          description: Null for FOUNDER; set for BRANCH_MANAGER and EMPLOYEE.
        unitId:
          type: [string, "null"]
          format: uuid
          description: Set for EMPLOYEE; null for FOUNDER and BRANCH_MANAGER.
        mustChangePassword:
          type: boolean
          description: True after an admin/temporary password reset; the client must force a change.
        canScore:
          type: boolean
          description: Granular grant to enter MANUAL/NUMERIC scores (FOUNDER always may).
        canReview:
          type: boolean
          description: Granular grant to review submissions (FOUNDER and a unit lead always may).
...
    RecoveryInviteRequest:
      type: object
      properties:
        expiresInMinutes:
          type: integer
          minimum: 1
          maximum: 43200
          default: 1440
          description: TTL for the recovery code. Defaults to 24 hours.
    RecoveryInviteResponse:
      type: object
      description: >-
        Returned once on issue. Carries the one-time recovery `code`; for a BRANCH_MANAGER target it
        also carries a one-time `temporaryPassword` (null for an EMPLOYEE, who has no password).
      required:
        - id
        - code
        - targetEmployeeId
        - role
        - expiresAt
        - createdAt
      properties:
        id:
          type: string
          format: uuid
        code:
          type: string
          description: Plaintext recovery code. Shown ONCE; only its hash is stored.
        targetEmployeeId:
          type: string
          format: uuid
        role:
          allOf:
            - $ref: "#/components/schemas/InvitableRole"
          description: The target user's role (EMPLOYEE or BRANCH_MANAGER); never FOUNDER.
        temporaryPassword:
          type: [string, "null"]
          description: One-time temporary password; set only for a BRANCH_MANAGER target, null otherwise.
        expiresAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
...
    EmployeePage:
      type: object
      required:
        - items
        - nextCursor
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/UserProfile"
        nextCursor:
          type: [string, "null"]
...
    SetPermissionsRequest:
      type: object
      required:
        - canScore
        - canReview
      properties:
        canScore:
          type: boolean
        canReview:
          type: boolean
...
    ManualGrade:
      type: string
      enum: [FULL, PARTIAL, ZERO]
      description: Multiplier of the item's points — FULL=1.0, PARTIAL=0.5, ZERO=0.0.
...
    ManualScoreRequest:
      type: object
      required:
        - grade
      properties:
        grade:
          $ref: "#/components/schemas/ManualGrade"
    NumericScoreRequest:
      type: object
      required:
        - value
      properties:
        value:
          type: number
    TaskScore:
      type: object
      required:
        - taskInstanceId
        - itemId
        - reviewerUserId
        - awardedPoints
      properties:
        taskInstanceId:
          type: string
          format: uuid
        itemId:
          type: string
          format: uuid
        reviewerUserId:
          type: string
          format: uuid
        grade:
          oneOf:
            - $ref: "#/components/schemas/ManualGrade"
            - type: "null"
        numericValue:
          type: [number, "null"]
        awardedPoints:
          type: integer
          description: Computed recommendation input; the system never computes pay.
    CreateNumericBandRequest:
      type: object
      required:
        - points
      properties:
        lowerBound:
          type: [number, "null"]
          description: Inclusive; null = -infinity.
        upperBound:
          type: [number, "null"]
          description: Exclusive; null = +infinity.
        points:
          type: integer
        label:
          type: [string, "null"]
          maxLength: 64
        sortOrder:
          type: integer
          default: 0
    NumericBand:
      type: object
      required:
        - id
        - itemId
        - points
        - sortOrder
      properties:
        id:
          type: string
          format: uuid
        itemId:
          type: string
          format: uuid
        lowerBound:
          type: [number, "null"]
        upperBound:
          type: [number, "null"]
        points:
          type: integer
        label:
          type: [string, "null"]
        sortOrder:
          type: integer
...
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          description: Stable machine-readable error code.
          example: INVALID_CREDENTIALS
        message:
          type: string
          description: Human-readable, non-sensitive description.
security:
  - bearerAuth: []